add script for historical server
